Moved Logic to service folder
